+++ src/App/App.bs.js
@@ -5,27 +5,27 @@ import * as React from "react";
 import * as Css from "@emotion/css";
 import * as CountrySelect from "./CountrySelect.bs.js";
 
-var container = Css.css({
+var app = Css.css({
+      display: "flex",
+      justifyContent: "center",
+      marginTop: "220px",
       fontFamily: "Arial, Helvetica, sans-serif",
       fontSize: "14px",
       lineHeight: "18px",
-      color: "#333333",
-      display: "flex",
-      justifyContent: "center",
-      marginTop: "220px"
+      color: "#333333"
     });
 
 var Styles = {
-  container: container
+  app: app
 };
 
 function App(Props) {
   var match = React.useState(function () {
-        return "us";
+        return "sg";
       });
   var setCountry = match[1];
   return React.createElement("div", {
-              className: container
+              className: app
             }, React.createElement(CountrySelect.make, {
                   country: match[0],
                   className: "country-select",
@@ -44,4 +44,4 @@ export {
   make ,
   
 }
-/* container Not a pure module */
+/* app Not a pure module */
