+++ src/App/App.res
@@ -1,21 +1,21 @@
 module Styles = {
   open Emotion
-  let container = css({
+  let app = css({
+    "display": "flex",
+    "justifyContent": "center",
+    "marginTop": "220px",
     "fontFamily": "Arial, Helvetica, sans-serif",
     "fontSize": "14px",
     "lineHeight": "18px",
     "color": "#333333",
-    "display": "flex",
-    "justifyContent": "center",
-    "marginTop": "220px",
   })
 }
 
 @react.component
 let make = () => {
-  let (country: string, setCountry) = React.useState(_ => "us")
+  let (country: string, setCountry) = React.useState(_ => "sg")
 
-  <div className={Styles.container}>
+  <div className={Styles.app}>
     <CountrySelect className="country-select" country onChange={value => setCountry(_ => value)} />
   </div>
 }
