+++ src/App/CountrySelect.res
@@ -1,37 +1,9 @@
 %%raw("import 'flag-icons/css/flag-icons.min.css'")
 %%raw("import './styles/CountrySelect.scss'")
 
-module SearchIcon = {
-  @module("./assets/search_icon.svg") @react.component
-  external make: (~className: string) => React.element = "ReactComponent"
-}
-
-module Styles = {
-  open Emotion
-  let control = css({
-    "minWidth": "230",
-    "margin": "8",
-  })
-  let menu = {"boxShadow": "inset 0 1px 0 rgba(0, 0, 0, 0.1)"}
-
-  let searchIcon = css({
-    "height": 13,
-    "width": 13,
-  })
-
-  // let dropdownIndicator = css({
-  //   "order": "-1",
-  // })
-}
-
-module SearchIconComponent = {
-  @react.component
-  let make = () => <div> <SearchIcon className={Styles.searchIcon} /> </div>
-}
-
 let getComponentsWithListRef = listRef => {
   let components: ReactSelect.components = {
-    dropdownIndicator: () => <SearchIconComponent />,
+    dropdownIndicator: () => <SearchIcon />,
     indicatorSeparator: () => React.null,
     menuList: props =>
       <CountrySelectMenu menuProps={props} height=160 itemSize=26 listRef={listRef} />,
@@ -44,6 +16,7 @@ let getComponentsWithListRef = listRef => {
 @react.component
 let make = (~country: string, ~className: string, ~onChange) => {
   let listRef = React.useRef(Js.Nullable.null)
+  let components = getComponentsWithListRef(listRef)
 
   let (options: array<Api.countryItem>, setOptions) = React.useState(_ => [])
   let (menuIsOpen, setMenuIsOpen) = React.useState(_ => false)
@@ -57,7 +30,6 @@ let make = (~country: string, ~className: string, ~onChange) => {
   }
 
   let currentCountry = Js.Array2.find(options, option => option.value === country)
-  let components = getComponentsWithListRef(listRef)
 
   React.useEffect0(() => {
     let _ = {
@@ -94,10 +66,11 @@ let make = (~country: string, ~className: string, ~onChange) => {
         switch Js.Nullable.toOption(listEl.props.itemData) {
         | None => ()
         | Some(itemData) =>
-          let index = Js.Array2.findIndex(options, option => option.value === itemData.value)
+          let currentIndex = Js.Array2.findIndex(options, option => option.value === itemData.value)
+          // fix scroll on move from start to end of a list, and backward
           let indexToScroll = switch key {
-          | "ArrowUp" => index === 0 ? Js.Array2.length(options) : index - 1
-          | _ => index === Js.Array2.length(options) - 1 ? 0 : index + 1
+          | "ArrowUp" => currentIndex === 0 ? Js.Array2.length(options) : currentIndex - 1
+          | _ => currentIndex === Js.Array2.length(options) - 1 ? 0 : currentIndex + 1
           }
           listEl.scrollToItem(. indexToScroll, "auto")
         }
@@ -106,34 +79,30 @@ let make = (~country: string, ~className: string, ~onChange) => {
   }
 
   <div className>
-    {switch Js.Array2.length(options) > 0 {
-    | false => React.null
-    | true =>
-      <DropDown
-        isOpen={menuIsOpen}
-        onClose={onToggleHandler}
-        target={<Button
-          text={switch currentCountry {
-          | Some(option) => option.label
-          | None => ""
-          }}
-          onClick={onToggleHandler}
-        />}>
-        <ReactSelect
-          value={currentCountry}
-          onChange={onChangeHandler}
-          options
-          placeholder="Search"
-          menuIsOpen
-          autoFocus={true}
-          controlShouldRenderValue={false}
-          classNamePrefix="--country-select"
-          components
-          escapeClearsValue={true}
-          onKeyDown
-          tabSelectsValue={true}
-        />
-      </DropDown>
-    }}
+    <DropDown
+      isOpen={menuIsOpen}
+      onClose={onToggleHandler}
+      target={<Button
+        text={switch currentCountry {
+        | Some(option) => option.label
+        | None => ""
+        }}
+        onClick={onToggleHandler}
+      />}>
+      <ReactSelect
+        value={currentCountry}
+        onChange={onChangeHandler}
+        options
+        placeholder="Search"
+        menuIsOpen
+        autoFocus={true}
+        controlShouldRenderValue={false}
+        classNamePrefix="--country-select"
+        components
+        escapeClearsValue={true}
+        onKeyDown
+        tabSelectsValue={true}
+      />
+    </DropDown>
   </div>
 }
