+++ src/App/SearchIcon.bs.js
@@ -1,0 +1,36 @@
+// Generated by ReScript, PLEASE EDIT WITH CARE
+
+import * as React from "react";
+import * as Css from "@emotion/css";
+import * as Search_iconSvg from "./assets/search_icon.svg";
+
+var make = Search_iconSvg.ReactComponent;
+
+var SearchIcon = {
+  make: make
+};
+
+var searchIcon = Css.css({
+      height: 13,
+      width: 13
+    });
+
+var Styles = {
+  searchIcon: searchIcon
+};
+
+function SearchIcon$1(Props) {
+  return React.createElement("div", undefined, React.createElement(make, {
+                  className: searchIcon
+                }));
+}
+
+var make$1 = SearchIcon$1;
+
+export {
+  SearchIcon ,
+  Styles ,
+  make$1 as make,
+  
+}
+/* make Not a pure module */
