+++ webpack.config.js
@@ -29,6 +29,11 @@ module.exports = {
           },
           {
             loader: 'file-loader',
+            options: {
+              name: '[name].[ext]',
+              outputPath: 'img',
+              useRelativePath: true,
+            },
           },
         ],
         type: 'javascript/auto',
